moving stuff to right locations

+++ scripts/watchdog.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 light_red='\e[1;91m%s\e[0m\n'                     
 light_green='\e[1;92m%s\e[0m\n'     
-camera_ip_address='192.168.1.13'
-locahost_ip_address='192.168.1.1'
+camera_ip_address='191.168.86.6'
+locahost_ip_address='191.168.86.1'
 ping_url='debian.org'
 restart_network=false
 restart_service=false
@@ -11,13 +11,12 @@ network_status=true
 check_network_status()
 {
   #Check eh0 and connectivity to IP camera 
-  network_on=0             
   ping -c 2 -q $camera_ip_address > /dev/null             
-  if [ "$?" -eq 0 ]; then                           
+  if [ "$?" -eq 0 ]; then                          
     printf "$light_green" "[ IP Camera Connected ]"
   else                                              
     printf "$light_red" "[ IP Camera Disconnected ]"
-    network_on=1    
+    return 1    
   fi
 
 
@@ -27,9 +26,9 @@ check_network_status()
     printf "$light_green" "[ WiFi Connected ]"
   else                                              
     printf "$light_red" "[ WiFi Disconnected ]"     
-    network_on=2
+    return 2
   fi
-  return $network_on
+  return 0
 }
 
 reset_ip_network()
@@ -47,7 +46,8 @@ reset_wifi_network()
 
 #retry to connect three times
 network_reset_happened=false
-for i in 1 2 3
+network_connected=false
+for i in 1 2 3 4 5
 do
   check_network_status
   network_on=$?
@@ -55,21 +55,20 @@ do
     printf "$light_green" "[ Resetting IP Camera... ]"
     reset_ip_network   
   elif [ $network_on == 2 ]; then
-    printf "$light_green" "[ Resetting WiFi Network... ]"
+    printf "$light_green" "[ Resetting WiFi Network - attempt $i ]"
     reset_wifi_network
     network_reset_happened=true
     #Wait for network to come back online
     sleep 60s
+
   else
     printf "$light_green" "[ Network Connected ]"
-    exit
+    network_connected=true
+    
   fi
 done
 
-if [ $network_reset_happened ]; then
+if [ $network_reset_happened ] && [ network_connected ]; then
   systemctl stop ngrok.service
   systemctl start ngrok.service
 fi
-
-#If the code comes here, it means the network was not 
-#connected even after three attemps and we will need to reboot.
